refactor(web): clear self-review suggestions — neutral wizard CSS + single bulk-config predicate (PR #1099)
Pays down the two SUGGESTION items from the self tech-review so the PR
carries no known tech-debt:

- Erli create-offer wizard no longer borrows Allegro's CSS namespace
  (`allegro-create-offer-wizard*` → neutral `create-offer-wizard*`). The
  classes carry no CSS today, so this is a rename-only, no-visual change
  that stops an Allegro-specific style tweak from silently affecting Erli.
- `allegroBulkConfigIsComplete` is no longer defined twice. The pure
  predicate moves into a component-free `allegro-bulk-config.schema.ts`
  (mirrors `erli-offer-fields.schema.ts`), is exported from the listings
  barrel, and the Allegro plugin imports it instead of re-implementing a
  private copy. The plugin keeps lazy-loading the section component; the
  predicate module has no component import, so no lazy chunk is pulled.

type-check ✅, eslint ✅, web suite ✅ 1563/1563.
